Refactor cli/outstanding.rs into modular structure
Break down the monolithic cli/outstanding.rs (3416 lines) into a cleaner
module structure with cohesive, single-purpose files.

## Changes

**Deleted:**
- `cli/outstanding.rs` (3416 lines) - monolithic file containing App and AppBuilder

**Created:**
- `cli/app.rs` (644 lines) - App struct, impl, and helper functions
- `cli/builder/` - AppBuilder submodule with 4 files:
  - `mod.rs` (186 lines) - Core struct, new(), Default, build(), parse()
  - `config.rs` (592 lines) - Configuration: context, topics, themes, templates, flags
  - `commands.rs` (585 lines) - Command registration: group, command, hooks
  - `execution.rs` (1500 lines) - Dispatch macro integration and execution

**Modified:**
- `cli/mod.rs` - Updated module declarations and re-exports
- `cli/group.rs` - Fixed import path for get_terminal_width

## Design decisions

- Unit tests remain colocated with the code they test (not in separate test files)
- Clear separation of concerns:
  - `config.rs`: How to configure the builder
  - `commands.rs`: How to register commands/handlers
  - `execution.rs`: How to run commands (dispatch, run, run_to_string)
- Named `execution.rs` (not `dispatch.rs`) to avoid confusion with existing
  `cli/dispatch.rs` which contains low-level dispatch primitives

All 580 tests pass.
